ARU-44 Moved constraints
