+++ src/main/java/com/amtinez/api/rest/users/validations/validators/UniqueUserEmailValidator.java
@@ -1,7 +1,7 @@
 package com.amtinez.api.rest.users.validations.validators;
 
-import com.amtinez.api.rest.users.annotations.UniqueUserEmail;
 import com.amtinez.api.rest.users.services.UserService;
+import com.amtinez.api.rest.users.validations.constraints.UniqueUserEmail;
 
 import javax.annotation.Resource;
 import javax.validation.ConstraintValidator;
